Converted whole project from double floating precision to single floating precision Single floating point precision will be necessary in the switch to GPU compute tasks

+++ LiteRay/Camera.swift
@@ -1,6 +1,6 @@
 //
 //  Camera.swift
-//  Trace
+//  LiteRay
 //
 //  Created by Matthew Dillard on 3/17/16.
 //  Copyright © 2016 Matthew Dillard. All rights reserved.
@@ -10,13 +10,13 @@ import Cocoa
 import simd
 
 public class Camera : NSObject, Translatable {
-	public var position: double3
-	var frustrum: (near: Double, far: Double)
-	var lookDirection: double3
-	var upDirection: double3
-	var FOV: Double
+	public var position: float3
+	var frustrum: (near: Float, far: Float)
+	var lookDirection: float3
+	var upDirection: float3
+	var FOV: Float
 	
-	public init(position pos:double3 = double3(), lookDir:double3 = double3(x:1,y:0,z:0), FOV: Double = 90.0, nearClip near:Double = 1.0, farClip far:Double = 1000.0, upDir:double3 = double3(x:0,y:1,z:0)) {
+	public init(position pos:float3 = float3(), lookDir:float3 = float3(x:1,y:0,z:0), FOV: Float = 90.0, nearClip near:Float = 1.0, farClip far:Float = 1000.0, upDir:float3 = float3(x:0,y:1,z:0)) {
 		position = pos
 		lookDirection = lookDir.unit
 		self.FOV = FOV
@@ -35,16 +35,16 @@ public class Camera : NSObject, Translatable {
 	/// has to work, capture draws the image flipped along the Y-axis.
 	public func capture(scene: Scene, size: NSSize, AntiAliasing: UInt) -> NSImage {
 		// calculate the screen dimensions given the FOV
-		let uWidth = tan((FOV / 2.0) * (M_PI / 180.0))
-		let vHeight = (Double(size.height) / Double(size.width)) * uWidth
+		let uWidth = tanf((FOV / 2.0) * (Float(M_PI) / 180.0))
+		let vHeight = (Float(size.height) / Float(size.width)) * uWidth
 		
 		// calculate the coordinate frame for screenspace
 		let X₀ = (lookDirection ⨯ upDirection).unit
 		let Y₀ = (X₀ ⨯ lookDirection).unit
 		
 		// compute the average width of a pixel represented in screenspace
-		let deltaX = (2 * uWidth * X₀) / Double(size.width)
-		let deltaY = (2 * vHeight * Y₀) / Double(size.height)
+		let deltaX = (2 * uWidth * X₀) / Float(size.width)
+		let deltaY = (2 * vHeight * Y₀) / Float(size.height)
 		
 		// grab the top left of the screenspace as the starting point for our image
 		var currentLeft = position + lookDirection - (X₀ * uWidth) + (Y₀ * vHeight)
@@ -53,9 +53,9 @@ public class Camera : NSObject, Translatable {
 		var pixels = [HDRColor]()
 		
 		// calculate the value of each pixel
-		for _ in 0..<Int(size.height) {
+		for _ in 0..<UInt(size.height) {
 			var pixelPosition = currentLeft
-			for _ in 0..<Int(size.width) {
+			for _ in 0..<UInt(size.width) {
 				pixels.append(getPixel(scene, SSPP: pixelPosition, AntiAliasing: AntiAliasing, dims: (deltaX, deltaY)))
 				
 				pixelPosition += deltaX
@@ -66,14 +66,14 @@ public class Camera : NSObject, Translatable {
 		return imageFromRGB32Bitmap(pixels, size: size)
 	}
 	
-	private func getPixel(scene: Scene, SSPP: double3, AntiAliasing: UInt, dims: (width: double3, height: double3)) -> HDRColor {
+	private func getPixel(scene: Scene, SSPP: float3, AntiAliasing: UInt, dims: (width: float3, height: float3)) -> HDRColor {
 		var pixel = HDRColor()
 		
 		// collect samples of the scene for this current pixel
 		for _ in 0...AntiAliasing {
 			// randomly generate offsets for the current subsample
-			let horiOffset = drand48()
-			let vertOffset = drand48()
+			let horiOffset = Float(drand48())
+			let vertOffset = Float(drand48())
 			
 			// get the subsample position and construct a ray from it
 			let subsample = SSPP + (dims.width * horiOffset) + (dims.height * vertOffset)
@@ -83,7 +83,7 @@ public class Camera : NSObject, Translatable {
 		}
 		
 		// return the normalized supersampled value
-		return pixel / Double(AntiAliasing);
+		return pixel / Float(AntiAliasing);
 	}
 	
 	private func trace(ray: Ray, scene: Scene, step: UInt, maxDepth: UInt, ignore: Shape? = nil) -> HDRColor {
@@ -108,7 +108,7 @@ public class Camera : NSObject, Translatable {
 		return getColor(scene, of: closest!, at: ray * zValue, from: -ray.d)
 	}
 	
-	private func getColor(scene: Scene, of shape: Shape, at point: double3, from: double3) -> HDRColor {
+	private func getColor(scene: Scene, of shape: Shape, at point: float3, from: float3) -> HDRColor {
 		// color independant of all other lighting conditions
 		let glow = shape.glow
