+++ LiteRay/Light.swift
@@ -1,6 +1,6 @@
 //
 //  Light.swift
-//  Trace
+//  LiteRay
 //
 //  Created by Matthew Dillard on 3/12/16.
 //  Copyright © 2016 Matthew Dillard. All rights reserved.
@@ -13,9 +13,9 @@ import simd
 public protocol Light {
 	var color: HDRColor { get set }
 	
-	func normalToLight(point: double3) -> double3
+	func normalToLight(point: float3) -> float3
 	
-	func illuminated(point: double3) -> Bool
+	func illuminated(point: float3) -> Bool
 	
-	func distance(point: double3) -> Double
+	func distance(point: float3) -> Float
 }
